did project architecture, added all typed of graphs
