+++ src/main/kotlin/graphs/primitives/Edge.kt
